requirements.txt
numpy>=1.26.0
pandas>=2.1.0
yfinance>=0.2.28
redis>=5.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
arch>=6.3.0
statsmodels>=0.14.0
scipy>=1.11.0
scikit-learn>=1.3.0
nolds>=0.5.2
ruptures>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
notebook>=7.0.0
asyncpg>=0.29.0
aiohttp>=3.9.0
websockets>=12.0
hmmlearn>=0.3.0
greenlet>=3.0.0
openai>=1.0.0
